tests: Refactor test_symbol_caching (now renamed test_caching)
